refactor(ProjectUsersRoleDropdown): absorb RoleSelect, add dirty prop, delete duplicate
ProjectUsersRoleSelect was a near-identical copy of ProjectUsersRoleDropdown
with only a dirty prop and highlight styles added. Consolidate into one component
and update ProjectUsersList and tests accordingly.
