+++ src/components/ProjectUsers/ProjectUsersRoleDropdown.vue
@@ -16,6 +16,10 @@ const props = defineProps({
   projectName: {
     type: String,
     required: true
+  },
+  dirty: {
+    type: Boolean,
+    default: false
   }
 })
 
@@ -37,42 +41,60 @@ defineExpose({ availableRoles })
 </script>
 
 <template>
-  <b-dropdown
+  <div
     class="project-users-role-dropdown"
-    no-caret
-    variant="body"
-    teleport-to="body"
-    toggle-class="project-users-role-dropdown__toggle"
+    :class="{ 'project-users-role-dropdown--dirty': dirty }"
   >
-    <template #button-content>
-      <div class="project-users-role-dropdown__content d-flex justify-content-between rounded-1">
-        <display-role :value="modelValue" />
-        <app-icon class="ms-2">
-          <i-ph-caret-down />
-        </app-icon>
-      </div>
-    </template>
-    <b-dropdown-item
-      v-for="role in availableRoles"
-      :key="role.value"
-      :active="role.value === modelValue"
-      @click="emit('update:modelValue', role.value)"
+    <b-dropdown
+      class="project-users-role-dropdown__dropdown"
+      no-caret
+      variant="body"
+      teleport-to="body"
+      toggle-class="project-users-role-dropdown__toggle"
     >
-      <display-role :value="role.value" />
-    </b-dropdown-item>
-  </b-dropdown>
+      <template #button-content>
+        <div class="project-users-role-dropdown__content d-flex justify-content-between rounded-1">
+          <display-role :value="modelValue" />
+          <app-icon class="ms-2">
+            <i-ph-caret-down />
+          </app-icon>
+        </div>
+      </template>
+      <display-role :value="modelValue" class="project-users-role-dropdown__value d-none" aria-hidden="true" />
+      <b-dropdown-item
+        v-for="role in availableRoles"
+        :key="role.value"
+        :active="role.value === modelValue"
+        @click="emit('update:modelValue', role.value)"
+      >
+        <display-role :value="role.value" />
+      </b-dropdown-item>
+    </b-dropdown>
+  </div>
 </template>
 
 <style scoped lang="scss">
 .project-users-role-dropdown {
-  border-radius: $btn-border-radius;
-  border: 1px solid $border-color;
+  &__dropdown {
+    border-radius: $btn-border-radius;
+    border: 1px solid $border-color;
+  }
+
   &__toggle {
     width: 9rem;
     justify-content: space-between;
   }
+
   :deep(.project-users-role-dropdown__content) {
     width: 11rem;
+  }
+
+  &--dirty &__dropdown {
+    border-color: var(--bs-primary);
+  }
+
+  &--dirty :deep(.project-users-role-dropdown__toggle) {
+    background-color: var(--bs-primary-bg-subtle) !important;
   }
 }
 </style>
