+++ tests/unit/specs/components/ProjectUsers/ProjectUsersList.spec.js
@@ -6,7 +6,7 @@ import ProjectUsersList from '@/components/ProjectUsers/ProjectUsersList.vue'
 import ProjectUsersActions from '@/components/ProjectUsers/ProjectUsersActions.vue'
 import ProjectUsersAdminPromotionModal from '@/components/ProjectUsers/ProjectUsersAdminPromotionModal.vue'
 import ProjectUsersCreateModal from '@/components/ProjectUsers/ProjectUsersCreateModal.vue'
-import ProjectUsersRoleSelect from '@/components/ProjectUsers/ProjectUsersRoleSelect.vue'
+import ProjectUsersRoleDropdown from '@/components/ProjectUsers/ProjectUsersRoleDropdown.vue'
 import EmptyState from '@/components/EmptyState/EmptyState.vue'
 import FormControlSearch from '@/components/Form/FormControl/FormControlSearch.vue'
 import { apiInstance as api } from '@/api/apiInstance.js'
@@ -93,9 +93,9 @@ describe('ProjectUsersList.vue', () => {
     expect(wrapper.findComponent(FormControlSearch).exists()).toBe(true)
   })
 
-  it('renders a ProjectUsersRoleSelect in each role cell', () => {
+  it('renders a ProjectUsersRoleDropdown in each role cell', () => {
     const wrapper = mountComponent()
-    expect(wrapper.findAllComponents(ProjectUsersRoleSelect)).toHaveLength(users.length)
+    expect(wrapper.findAllComponents(ProjectUsersRoleDropdown)).toHaveLength(users.length)
   })
 
   it('renders a ProjectUsersActions in each row', () => {
@@ -124,23 +124,23 @@ describe('ProjectUsersList.vue', () => {
 
   it('passes user.role as modelValue when no pending change', () => {
     const wrapper = mountComponent()
-    expect(wrapper.findAllComponents(ProjectUsersRoleSelect)[0].props('modelValue')).toBe(users[0].role)
+    expect(wrapper.findAllComponents(ProjectUsersRoleDropdown)[0].props('modelValue')).toBe(users[0].role)
   })
 
   it('passes dirty=false when no pending change', () => {
     const wrapper = mountComponent()
-    expect(wrapper.findAllComponents(ProjectUsersRoleSelect)[0].props('dirty')).toBe(false)
+    expect(wrapper.findAllComponents(ProjectUsersRoleDropdown)[0].props('dirty')).toBe(false)
   })
 
   it('updates pendingChanges when update:modelValue is emitted by a role select', async () => {
     const wrapper = mountComponent()
-    await wrapper.findAllComponents(ProjectUsersRoleSelect)[0].vm.$emit('update:modelValue', 'PROJECT_MEMBER')
+    await wrapper.findAllComponents(ProjectUsersRoleDropdown)[0].vm.$emit('update:modelValue', 'PROJECT_MEMBER')
     expect(wrapper.vm.pendingChanges[users[0].name]).toBe('PROJECT_MEMBER')
   })
 
   it('passes dirty=true and updated modelValue after a role change', async () => {
     const wrapper = mountComponent()
-    const roleSelect = wrapper.findAllComponents(ProjectUsersRoleSelect)[0]
+    const roleSelect = wrapper.findAllComponents(ProjectUsersRoleDropdown)[0]
     await roleSelect.vm.$emit('update:modelValue', 'PROJECT_MEMBER')
     await wrapper.vm.$nextTick()
     expect(roleSelect.props('dirty')).toBe(true)
@@ -149,7 +149,7 @@ describe('ProjectUsersList.vue', () => {
 
   it('removes entry from pendingChanges when role reverts to original', async () => {
     const wrapper = mountComponent()
-    const roleSelect = wrapper.findAllComponents(ProjectUsersRoleSelect)[0]
+    const roleSelect = wrapper.findAllComponents(ProjectUsersRoleDropdown)[0]
     await roleSelect.vm.$emit('update:modelValue', 'PROJECT_MEMBER')
     await roleSelect.vm.$emit('update:modelValue', users[0].role)
     expect(wrapper.vm.pendingChanges[users[0].name]).toBeUndefined()
@@ -164,7 +164,7 @@ describe('ProjectUsersList.vue', () => {
 
   it('shows sticky bar when there are pending changes', async () => {
     const wrapper = mountComponent()
-    await wrapper.findAllComponents(ProjectUsersRoleSelect)[0].vm.$emit('update:modelValue', 'PROJECT_MEMBER')
+    await wrapper.findAllComponents(ProjectUsersRoleDropdown)[0].vm.$emit('update:modelValue', 'PROJECT_MEMBER')
     await wrapper.vm.$nextTick()
     expect(wrapper.find('.project-users-list__sticky-bar').exists()).toBe(true)
   })
@@ -174,7 +174,7 @@ describe('ProjectUsersList.vue', () => {
   it('saveRoles calls saveProjectPolicy for each pending change', async () => {
     api.saveProjectPolicy.mockResolvedValue(undefined)
     const wrapper = mountComponent()
-    await wrapper.findAllComponents(ProjectUsersRoleSelect)[0].vm.$emit('update:modelValue', 'PROJECT_MEMBER')
+    await wrapper.findAllComponents(ProjectUsersRoleDropdown)[0].vm.$emit('update:modelValue', 'PROJECT_MEMBER')
     await wrapper.vm.saveRoles()
     await flushPromises()
     expect(api.saveProjectPolicy).toHaveBeenCalledWith('default', projectName, {
@@ -186,7 +186,7 @@ describe('ProjectUsersList.vue', () => {
   it('saveRoles updates localUsers and clears pendingChanges on success', async () => {
     api.saveProjectPolicy.mockResolvedValue(undefined)
     const wrapper = mountComponent()
-    await wrapper.findAllComponents(ProjectUsersRoleSelect)[0].vm.$emit('update:modelValue', 'PROJECT_MEMBER')
+    await wrapper.findAllComponents(ProjectUsersRoleDropdown)[0].vm.$emit('update:modelValue', 'PROJECT_MEMBER')
     await wrapper.vm.saveRoles()
     await flushPromises()
     expect(wrapper.vm.pendingChanges).toEqual({})
@@ -196,7 +196,7 @@ describe('ProjectUsersList.vue', () => {
   it('saveRoles shows success toast on success', async () => {
     api.saveProjectPolicy.mockResolvedValue(undefined)
     const wrapper = mountComponent()
-    await wrapper.findAllComponents(ProjectUsersRoleSelect)[0].vm.$emit('update:modelValue', 'PROJECT_MEMBER')
+    await wrapper.findAllComponents(ProjectUsersRoleDropdown)[0].vm.$emit('update:modelValue', 'PROJECT_MEMBER')
     await wrapper.vm.saveRoles()
     await flushPromises()
     expect(mockToast.success).toHaveBeenCalledOnce()
@@ -205,7 +205,7 @@ describe('ProjectUsersList.vue', () => {
   it('saveRoles clears pendingChanges and shows error toast on API failure', async () => {
     api.saveProjectPolicy.mockRejectedValue(new Error('forbidden'))
     const wrapper = mountComponent()
-    await wrapper.findAllComponents(ProjectUsersRoleSelect)[0].vm.$emit('update:modelValue', 'PROJECT_MEMBER')
+    await wrapper.findAllComponents(ProjectUsersRoleDropdown)[0].vm.$emit('update:modelValue', 'PROJECT_MEMBER')
     await wrapper.vm.saveRoles()
     await flushPromises()
     expect(wrapper.vm.pendingChanges).toEqual({})
@@ -216,7 +216,7 @@ describe('ProjectUsersList.vue', () => {
 
   it('cancelChanges clears all pending changes', async () => {
     const wrapper = mountComponent()
-    await wrapper.findAllComponents(ProjectUsersRoleSelect)[0].vm.$emit('update:modelValue', 'PROJECT_MEMBER')
+    await wrapper.findAllComponents(ProjectUsersRoleDropdown)[0].vm.$emit('update:modelValue', 'PROJECT_MEMBER')
     wrapper.vm.cancelChanges()
     await wrapper.vm.$nextTick()
     expect(wrapper.vm.pendingChanges).toEqual({})
@@ -227,7 +227,7 @@ describe('ProjectUsersList.vue', () => {
   it('onSaveClicked opens admin modal when a pending change promotes to an admin role', async () => {
     // users[0] is PROJECT_EDITOR; promoting to PROJECT_ADMIN triggers the modal
     const wrapper = mountComponent()
-    await wrapper.findAllComponents(ProjectUsersRoleSelect)[0].vm.$emit('update:modelValue', 'PROJECT_ADMIN')
+    await wrapper.findAllComponents(ProjectUsersRoleDropdown)[0].vm.$emit('update:modelValue', 'PROJECT_ADMIN')
     wrapper.vm.onSaveClicked()
     await wrapper.vm.$nextTick()
     expect(wrapper.vm.showAdminModal).toBe(true)
@@ -237,7 +237,7 @@ describe('ProjectUsersList.vue', () => {
     api.saveProjectPolicy.mockResolvedValue(undefined)
     // users[1] is DOMAIN_ADMIN; changing to PROJECT_MEMBER is a demotion — no modal
     const wrapper = mountComponent()
-    await wrapper.findAllComponents(ProjectUsersRoleSelect)[1].vm.$emit('update:modelValue', 'PROJECT_MEMBER')
+    await wrapper.findAllComponents(ProjectUsersRoleDropdown)[1].vm.$emit('update:modelValue', 'PROJECT_MEMBER')
     wrapper.vm.onSaveClicked()
     await flushPromises()
     expect(api.saveProjectPolicy).toHaveBeenCalled()
